extract some more helpers
